+++ plom/server/plomServer/routesID.py
@@ -343,7 +343,6 @@ class IDHandler:
         elif what == 403:
             raise web.HTTPForbidden(reason=msg)
         else:
-            # catch all that should not happen.
             raise web.HTTPInternalServerError(reason=msg)
 
     # @routes.put("/ID/{papernum}")
@@ -354,6 +353,10 @@ class IDHandler:
         Only "manager" can perform this action.  Typical client IDing
         would call func:`IdentifyPaperTask` instead.
 
+        By passing empty `sid` and `sname`, this API can be used to
+        unidentify a paper.  This feature may move to its own API call
+        in the future.
+
         Returns:
             400: not manager.
             404: papernum not found, or other data errors.
@@ -362,6 +365,13 @@ class IDHandler:
         if not data["user"] == "manager":
             raise web.HTTPBadRequest(reason="Not manager")
         papernum = request.match_info["papernum"]
+
+        # special feature to unidentify: move elsewhere?
+        if not data["sid"] and not data["sname"]:
+            if self.server.DB.remove_id_from_paper(papernum):
+                return web.Response(status=200)
+            raise web.HTTPNotFound(reason=f"Did not find papernum {papernum}")
+
         r, what, msg = self.server.id_paper(papernum, "HAL", data["sid"], data["sname"])
         if r:
             return web.Response(status=200)
@@ -370,7 +380,6 @@ class IDHandler:
         elif what == 404:
             raise web.HTTPNotFound(reason=msg)
         else:
-            # catch all that should not happen.
             raise web.HTTPInternalServerError(reason=msg)
 
     # @routes.delete("/ID/tasks/{task}")